requirements.txt
datasets==3.0.1
fitz==0.0.1.dev2
langchain==0.3.7
langchain_chroma==0.1.4
langchain_community==0.3.7
langchain_core==0.3.17
langchain_groq==0.2.1
langchain_huggingface==0.1.2
langchain_openai==0.2.8
langchain_text_splitters==0.3.2
langgraph==0.2.46
keyring>=24.0.0
pydantic==2.9.2
python-dotenv>=1.0.0
PyYAML==6.0.2
ragas==0.2.5
setuptools==68.0.0
streamlit==1.38.0
torch==2.3.1
transformers==4.40.2
timescale-vector
typing_extensions==4.12.2
rank_bm25
azure-ai-formrecognizer
azure-ai-formrecognizer>=3.2.0
matplotlib>=3.5.0
Pillow>=9.0.0
PyMuPDF>=1.19.0
python-frontmatter
langfuse
timescale_vector